make real_world_currency_per_point a float
then it will be easier to work with. for example, taking the reciprocal like 1/rwc_per_point. Decimal is more restrictive, whereas float can work with everything. i dont think we have to worry about decimal junk, because this value will practically always be written as a literal, not the result of an arithmetic operation