revert channels to pre-1.0
Because of this issue: https://github.com/django/channels/issues/498, we don't want to upgrade to Channels 1.0+ because that means Windows users would need to install pypiwin32. This can get complicated because when they do 'pip freeze' it will get written to their requirements file then fail to install on Heroku. Also, it seems that even though we restrict pypiwin32 install to Windows users with a directive in requirements.txt, at least 1 user on Mac got it installed anyway. Until we upgrate to Channels 1.0+, we won't be ablet to do local development with HTTPS (for MTurk sandbox)